Move each viewmodel to ui package.
