+++ app/src/main/java/com/taufik/themovieshow/ui/discover/DiscoverMovieFragment.kt
@@ -16,7 +16,7 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.taufik.themovieshow.R
-import com.taufik.themovieshow.data.viewmodel.movie.MovieViewModel
+import com.taufik.themovieshow.ui.main.movie.viewmodel.MovieViewModel
 import com.taufik.themovieshow.databinding.FragmentDiscoverMovieBinding
 import com.taufik.themovieshow.ui.main.movie.adapter.DiscoverMovieAdapter
