+++ app/src/main/java/com/taufik/themovieshow/ui/main/favorite/fragment/FavoriteMovieFragment.kt
@@ -15,7 +15,7 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.taufik.themovieshow.R
 import com.taufik.themovieshow.data.local.entity.FavoriteMovie
-import com.taufik.themovieshow.data.viewmodel.movie.FavoriteMovieViewModel
+import com.taufik.themovieshow.ui.main.movie.viewmodel.FavoriteMovieViewModel
 import com.taufik.themovieshow.databinding.FragmentFavoriteMovieBinding
 import com.taufik.themovieshow.ui.main.movie.adapter.MovieAdapter
