+++ app/src/main/java/com/taufik/themovieshow/ui/main/favorite/fragment/FavoriteTvShowsFragment.kt
@@ -15,7 +15,7 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.taufik.themovieshow.R
 import com.taufik.themovieshow.data.local.entity.FavoriteTvShow
-import com.taufik.themovieshow.data.viewmodel.tvshow.FavoriteTvShowViewModel
+import com.taufik.themovieshow.ui.main.tvshow.viewmodel.FavoriteTvShowViewModel
 import com.taufik.themovieshow.databinding.FragmentFavoriteTvShowsBinding
 import com.taufik.themovieshow.ui.main.tvshow.adapter.TvShowsAdapter
