+++ app/src/main/java/com/taufik/themovieshow/ui/main/movie/fragment/MovieUpcomingFragment.kt
@@ -8,7 +8,7 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.taufik.themovieshow.data.viewmodel.movie.MovieViewModel
+import com.taufik.themovieshow.ui.main.movie.viewmodel.MovieViewModel
 import com.taufik.themovieshow.databinding.FragmentMovieUpcomingBinding
 import com.taufik.themovieshow.ui.main.movie.adapter.MovieAdapter
