+++ app/src/main/java/com/taufik/themovieshow/ui/main/movie/viewmodel/MovieViewModel.kt
@@ -1,4 +1,4 @@
-package com.taufik.themovieshow.data.viewmodel.movie
+package com.taufik.themovieshow.ui.main.movie.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
